dataset_pipeline/requirements.txt
google-genai
python-dotenv
pillow
PyYAML
tqdm